backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
chromadb==0.5.0
sentence-transformers==2.3.1
google-generativeai==0.3.2
python-dotenv==1.0.0
